Renaming of config directories for clarity of dags/pipeline_plugin config more explicit.
