+++ dags/hdx_adm_dags.py
@@ -3,7 +3,7 @@ from airflow.operators import BashOperator
 from datetime import datetime, timedelta
 from airflow.operators.pipeline_plugin import HDXExtractOperator, HDXAdm0TransformOperator, HDXAdm1TransformOperator
 
-from config import config
+from country_config import config
 
 def create_hdx_adm_dag(countries, schedule_interval, catchup, default_args):
     dag = DAG(f"hdx_adm", schedule_interval=schedule_interval, catchup=catchup, default_args=default_args)
