--- plugins/pipeline_plugin/config/__init__.py
@@ -1,1 +1,0 @@
-from pipeline_plugin.config.config_parser import config
