+++ plugins/pipeline_plugin/pipeline_config/config.develop.yaml
