+++ plugins/pipeline_plugin/pipeline_config/config.staging.yaml
