+++ plugins/pipeline_plugin/utils/google_cloud_storage.py
@@ -1,5 +1,5 @@
 from google.cloud import storage
-from pipeline_plugin.config import config
+from pipeline_plugin.pipeline_config import config
 from pathlib import Path
